requirements.txt
alembic==1.13.3
Werkzeug==2.2.2
Flask==2.2.5
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
pandas==2.2.1
selenium==4.19.0
wtforms==3.0.1
Flask-WTF==1.1.1
requests==2.31.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Flask-Login==0.6.3
gunicorn
email-validator
openpyxl
psycopg2-binary
python-dotenv
sib-api-v3-sdk==7.6.0
apscheduler
cryptography
fuzzywuzzy==0.18.0
python-dotenv==1.0.0
python-Levenshtein==0.21.0
